requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.9

# AWS and cloud services
boto3==1.34.0
botocore==1.34.0

# Database and vector storage
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.41
alembic>=1.16.4
pgvector>=0.4.1

# PDF processing and text extraction
pymupdf>=1.26.3
pdfplumber>=0.11.7
pypdf>=5.9.0

# ML and embeddings
openai>=1.97.1
sentence-transformers>=5.0.0
scikit-learn>=1.7.1
numpy>=2.3.2

# Data processing
pandas>=2.3.1
pydantic>=2.11.7
pyyaml>=6.0.2

# Text processing and search
rank-bm25>=0.2.2
nltk>=3.9.1

# HTTP and web scraping
requests>=2.32.4
beautifulsoup4>=4.13.4
httpx>=0.28.1

# Logging and monitoring
structlog>=25.4.0
prometheus-client>=0.22.1

# Development and testing
pytest>=8.4.1
pytest-asyncio>=1.1.0
black>=25.1.0
flake8>=7.3.0